public_html/api/requirements-cpanel.txt
# Django and REST Framework - Compatible with Python 3.12.11
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0

# Image processing
Pillow==10.4.0

# Environment variables
python-decouple==3.8

# Database - Using PyMySQL for better compatibility
PyMySQL==1.1.1

# Static files serving for production
whitenoise==6.8.2

# WSGI server for production
gunicorn==23.0.0

# Security
cryptography==43.0.3

# Utilities
python-dateutil==2.9.0

# PDF generation
reportlab==4.2.5
jdatetime==5.0.0
arabic-reshaper==3.0.0
python-bidi==0.4.2

# WebSocket support
channels==4.1.0
daphne==4.1.2

# Jalali date support
django-jalali==6.0.1

# TinyMCE editor
django-tinymce==4.1.0

# Additional dependencies
requests==2.32.3
asgiref==3.8.1
sqlparse==0.5.2
pytz==2024.2